ovnkube in dpu host mode: advanced gateway detection
This patch introduces a new way to detect the gateway interface in the case
ovnkube is running in DPU-HOST mode.

Introduce the argument derive-from-mgmt-port, if this is specified as the
gateway interface, we will identify the physical function of the device
used as an management port accelerated interface.

Cherry-picked from upstream commit a6cac97ae
Signed-off-by: Alin Gabriel Serdean <aserdean@nvidia.com>